Create CSVeditor
Sure! Here’s the complete description on how to install and use CSVeditor.py:

CSV Editor

CSV Editor is a simple Python script that enables users to manipulate CSV files effortlessly. This script allows users to remove specific columns from a CSV file. Users are prompted to input the location of the CSV file they want to modify, specify the name of the column to be deleted, and designate the location to save the edited CSV file. The script verifies the existence of the specified column for deletion in the CSV file, then removes it from both the header and the data, and saves the modified data to a new CSV file.

Installation:

	1.	Make sure you have Python 3.x installed on your computer. If not, you can download it from the official Python website.
	2.	Download the CSVeditor.py file to your computer.
	3.	Open a terminal or command prompt, then navigate to the directory where you have saved the CSVeditor.py file.
	4.	Run the script using the following command:

python CSVeditor.py



Usage:

	1.	After running the script, you will be prompted to enter the location of the CSV file you want to edit. Make sure to provide the full path (e.g., /path/to/input_file.csv).
	2.	Next, you will see a list of columns available in the CSV file.
	3.	Enter the name of the column you want to delete from the CSV file.
	4.	The script will process the deletion of the requested column and then ask you to specify the location where you want to save the edited CSV file.
	5.	Once the process is complete, you will receive a message confirming that the column has been deleted and the edited CSV file has been successfully saved.

Note: Ensure you have adequate permissions to read the input file and write to the output file.

If you have any questions or need further assistance, feel free to ask!